Add files via upload
I am proud to have completed the "A.I. for Everyone" course on Coursera and obtained a certificate of completion. Through this course, I gained an understanding of the fundamentals of artificial intelligence (AI), its applications, and limitations. I also learned about various AI techniques, such as machine learning, and how they can be used in real-world scenarios.

This course has equipped me with the knowledge and tools to make informed decisions about the impact of AI on our world, and I am excited to continue learning about this rapidly evolving field.